requirements.txt
fastapi
uvicorn
sqlalchemy
pymupdf
python-multipart
langchain-core
chromadb
langchain-community
langchain-chroma
langchain-huggingface
langchain-ollama
sentence-transformers